Unauthorized 403 page

+++ app/Http/Controllers/Api/CommentController.php
@@ -9,7 +9,7 @@ use Illuminate\Http\Request;
 class CommentController extends Controller
 {
     public function destroy(Comment $comment){
-//        $this->authorize('delete', $comment);
+        $this->authorize('delete', $comment);
         $comment->delete();
         return back();
     }
